+++ src/main/resources/schema.sql
@@ -2,6 +2,6 @@ create table Advertiser
 (
    Advertiser_Name VARCHAR(255) not null,
    Contact_Name varchar(255) not null,
-   Credit_Limit INTEGER not null,
+   Credit_Limit FLOAT not null,
    primary key(Advertiser_Name)
 );
